requirements.txt
# AGI Autonomous Core V6.1 - Requirements

# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Optional but recommended
aiohttp>=3.9.0

# For development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
